requirements.txt
python-json-logger==2.0.2
python-lambda-local==0.1.12
